+++ HISTORY.md
@@ -4,7 +4,7 @@
 * **New Features:**
     * passing a list to the ASK command in sensing presents a menu to the user
     * formatting a list of texts displays it as chat-history in an ASK menu
-    * ASK nothing or a falsy value terminates the thread currently displaying a question
+    * ASK nothing or a falsy value terminates all threads currently displaying a question or waiting to ask one
     * new "Menus" library
     * export script (including dependencies) via its context menu
     * export / import sprite-local custom block definitions from the palette
@@ -39,6 +39,9 @@
     * updated the API documentation for "setTranslation"
 * **Translation Updates:**
     * German
+
+### 2022-04-20
+* threads: terminate all threads waiting to display a question on ASKing a falsy value
 
 ### 2022-04-19
 * threads: ASK nothing or a falsy value terminates the thread currently displaying a question